backend/requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
Django==3.2.9
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-js-asset==1.2.2
djangorestframework==3.12.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
lxml==4.6.4
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pylint==2.12.1
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
